+++ BelofteCheck/Controllers/WettenController.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
-using System.Data.SqlClient;
 using System.Linq;
-using System.Net;
-using System.Web;
 using System.Web.Mvc;
-using BelofteCheck;
 using BelofteCheck.ViewModels;
 
 namespace BelofteCheck.Controllers
@@ -92,6 +87,14 @@ namespace BelofteCheck.Controllers
 
             List<WetObject> q = query.ToList();
 
+            if (q == null)
+            {
+                TempData["BCmessage"] = "Wet ID " + WetID.Trim() + " is niet gevonden";
+                TempData["BCerrorlevel"] = wettenVM.MessageSection.Warning;
+
+                return RedirectToAction("Error");
+            }
+
 
             if (q[0].OnderwerpID == "<geen>")
             {                  
@@ -108,7 +111,12 @@ namespace BelofteCheck.Controllers
         // GET: Wetten/Create
         public ActionResult Create()
         {
-            return View(new Wetten());
+            WettenVM wettenVM = new WettenVM();
+            string title = "Nieuw";
+            string level = wettenVM.MessageSection.Info;
+            string msg = "Vul de gegevens voor de nieuwe wet in en selecteer AANMAKEN";
+            wettenVM.MessageSection.SetMessage(title, level, msg);
+            return View(wettenVM);
         }
 
         // POST: Wetten/Create
@@ -116,48 +124,114 @@ namespace BelofteCheck.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "WetID,WetNaam,WetOmschrijving,WetLink")] Wetten wetten)
+        public ActionResult Create([Bind(Include = "WetID,WetNaam,WetOmschrijving,WetLink")] WettenVM wettenVM)
         {
+            string title = "Nieuw";
+
             if (ModelState.IsValid)
             {
+                Wetten wetten = new Wetten();
+                wetten.Fill(wettenVM); 
                 db.Wetten.Add(wetten);
                 db.SaveChanges();
+                TempData["BCmessage"] = "Wet " + wettenVM.WetNaam.Trim() + " is nu aangemaakt";
+                TempData["BCerrorlevel"] = wettenVM.MessageSection.Info;
+                
                 return RedirectToAction("Index");
             }
 
-            return View(wetten);
+            string level = wettenVM.MessageSection.Error;
+            string msg = "ERROR - Wet " + wettenVM.WetNaam.Trim() + " is NIET aangemaakt";
+            wettenVM.MessageSection.SetMessage(title, level, msg);
+            return View(wettenVM);
             
         }
 
         // GET: Wetten/Edit/5
         public ActionResult Edit(string WetID)
         {
+            WettenVM wettenVM = new WettenVM();
+            string title = "Bewerken";
+            string level = wettenVM.MessageSection.Info;
+            string msg = "Bewerk deze wet en/of de gekoppelde onderwerpen en selecteer OPSLAAN";
+
             if (WetID == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                TempData["BCmessage"] = "Specificeer een geldige Wet ID!";
+                TempData["BCerrorlevel"] = wettenVM.MessageSection.Warning;
+
+                return RedirectToAction("Error");
             }
-            Wetten wetten = db.Wetten.Find(WetID);
-            if (wetten == null)
+            // Perform outer join form Wetten, Wetsccope, Onderwerpen
+
+            var query = from w in db.Wetten
+                        where w.WetID == WetID
+                        join s in db.WetScope on w.WetID equals s.WetID into ljoin1
+                        from lj1 in ljoin1.DefaultIfEmpty()
+                        join o in db.Onderwerpen on lj1.OnderwerpID equals o.OnderwerpID into ljoin2
+                        from lj2 in ljoin2.DefaultIfEmpty()
+                        select new WetObject
+                        {
+                            WetID = w.WetID,
+                            WetNaam = w.WetNaam,
+                            WetOmschrijving = w.WetOmschrijving,
+                            WetLink = w.WetLink,
+                            OnderwerpID = lj1 == null ? "<geen>" : lj1.OnderwerpID,
+                            Toelichting = lj1 == null ? "nvt" : lj1.Toelichting,
+                            Omschrijving = lj2 == null ? "nvt" : lj2.Omschrijving
+                        }
+                        ;
+
+            List<WetObject> q = query.ToList();
+
+            if (q == null)
             {
-                return HttpNotFound();
+                TempData["BCmessage"] = "Wet ID " + WetID.Trim() + " is niet gevonden";
+                TempData["BCerrorlevel"] = wettenVM.MessageSection.Warning;
+
+                return RedirectToAction("Error");
             }
-            return View(wetten);
+
+
+            if (q[0].OnderwerpID == "<geen>")
+            {
+                msg = "Deze wet heeft geen gekoppelde onderwerpen. Vink er minimaal één aan!";
+                level = "W";
+            }
+
+            wettenVM.Fill(q);
+            wettenVM.MessageSection.SetMessage(title, level, msg);
+
+            return View(wettenVM);
         }
+    
 
         // POST: Wetten/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "WetID,WetNaam,WetOmschrijving,WetLink")] Wetten wetten)
+        public ActionResult Edit([Bind(Include = "WetID,WetNaam,WetOmschrijving,WetLink")] WettenVM wettenVM)
         {
+            string title = "Bewerken";
+
             if (ModelState.IsValid)
             {
+                Wetten wetten = new Wetten();
+                wetten.Fill(wettenVM);
                 db.Entry(wetten).State = EntityState.Modified;
                 db.SaveChanges();
+                TempData["BCmessage"] = "Wet " + wettenVM.WetNaam.Trim() + " is gewijzigd";
+                TempData["BCerrorlevel"] = wettenVM.MessageSection.Info;
+
                 return RedirectToAction("Index");
             }
-            return View(wetten);
+
+            string level = wettenVM.MessageSection.Error;
+            string msg = "ERROR - Wet " + wettenVM.WetNaam.Trim() + " is NIET gewijzigd";
+            wettenVM.MessageSection.SetMessage(title, level, msg);
+            return View(wettenVM);
+            
         }
 
         // GET: Wetten/Delete/5
@@ -197,6 +271,14 @@ namespace BelofteCheck.Controllers
 
             List<WetObject> q = query.ToList();
 
+            if (q == null)
+            {
+                TempData["BCmessage"] = "Wet ID " + WetID.Trim() + " is niet gevonden";
+                TempData["BCerrorlevel"] = wettenVM.MessageSection.Warning;
+
+                return RedirectToAction("Error");
+            }
+
 
             if (q[0].OnderwerpID == "<geen>")
             {
@@ -225,6 +307,15 @@ namespace BelofteCheck.Controllers
             }
             
             Wetten wetten = db.Wetten.Find(WetID);
+
+            if (wetten == null)
+            {
+                TempData["BCmessage"] = "Wet ID " + WetID.Trim() + " is niet gevonden";
+                TempData["BCerrorlevel"] = wettenVM.MessageSection.Warning;
+
+                return RedirectToAction("Error");
+            }
+
             db.Wetten.Remove(wetten);
             db.SaveChanges();
